format ut dir tree
