+++ tests/test_models/test_task_modules/test_anchor/test_anchor_3d_generator.py
